Add leetcode problem [55] code
Solution idea. The first instinct was to model the jumps and
let the indices jump around in the array as far as they come.
Upon closer looking only arrays with 0s pose a problem, since
in all other arrays one can simply iterate through the array.

Exception handling with 0s at various tricky spots was the major
issue in solving the problem